+++ python/run.sh
@@ -1,0 +1,7 @@
+#!/bin/sh
+
+docker build --tag control-libraries/python/test .
+
+docker run -it --rm \
+  --name control-libraries-python-test \
+  control-libraries/python/test